Refactor structure
Replace duplicate code

Move argument processing to utils

Implement better error handling

Move variables and constants to the top
